Data.Vector.Fixed

Description

Generic API for vectors with fixed length.

For encoding of vector size library uses Peano naturals defined in the library. At come point in the future it would make sense to switch to new GHC type level numerals.

Common pitfalls

Library provide instances for tuples. But there's a catch. Tuples are monomorphic in element type. Let consider 2-tuple (Int,Int). Vector type v is (,) Int and only allowed element type is Int. Because of that we cannot change element type and following code will fail:

>>> map (== 1) ((1,2) :: (Int,Int))

<interactive>:3:1:
    Couldn't match type `Int' with `Bool'
    In the expression: F.map (== 1) ((1, 2) :: (Int, Int))
    In an equation for `it': it = map (== 1) ((1, 2) :: (Int, Int))

To make it work we need to change vector type as well. Functions from module Data.Vector.Fixed.Generic provide this functionality.

>>> map (== 1) ((1,2) :: (Int,Int)) :: (Bool,Bool)
(True,False)

Vector type class

Vector size

type family Dim (v :: * -> *) :: Nat

Size of vector expressed as type-level natural.

Type class

class Arity (Dim v) => Vector v a where

Type class for vectors with fixed length. Instance should provide two functions: one to create vector and another for vector deconstruction. They must obey following law:

inspect v construct = v

For example instance for 2D vectors could be written as:

data V2 a = V2 a a

type instance V2 = 2
instance Vector V2 a where
  construct                = Fun V2
  inspect (V2 a b) (Fun f) = f a b

Minimal complete definition

construct, inspect

Methods

construct :: Fun (Peano (Dim v)) a (v a)

N-ary function for creation of vectors.

inspect :: v a -> Fun (Peano (Dim v)) a b -> b

Deconstruction of vector.

basicIndex :: v a -> Int -> a

Optional more efficient implementation of indexing. Shouldn't be used directly, use ! instead.

class (Vector (v n) a, Dim (v n) ~ n) => VectorN v n a

Vector parametrized by length. In ideal world it should be:

forall n. (Arity n, Vector (v n) a, Dim (v n) ~ n) => VectorN v a

Alas polymorphic constraints aren't allowed in haskell.

type Arity n = (ArityPeano (Peano n), KnownNat n, Peano (n + 1) ~ 'S (Peano n))

Type class for type level number for which we can defined operations over N-ary functions.

newtype Fun n a b

Newtype wrapper which is used to make Fn injective. It's also a reader monad.

Constructors

Fun

Fields

unFun :: Fn n a b

length :: forall v a. KnownNat (Dim v) => v a -> Int

Length of vector. Function doesn't evaluate its argument.

Constructors

There are several ways to construct fixed vectors except using their constructor if it's available. For small ones it's possible to use functions mk1, mk2, etc.

>>> mk3 'a' 'b' 'c' :: (Char,Char,Char)
('a','b','c')

Alternatively one could use mkN. See its documentation for examples

Another option is to create tuple and convert it to desired vector type. For example:

v = convert (x,y,z)

It will work on if type of v is know from elsewhere. Same trick could be used to pattern match on the vector with opaque representation using view patterns

function :: Vec N3 Double -> ...
function (convert -> (x,y,z)) = ...

Constructors for vectors with small dimensions.

mk0 :: (Vector v a, Dim v ~ 0) => v a

mk1 :: (Vector v a, Dim v ~ 1) => a -> v a

mk2 :: (Vector v a, Dim v ~ 2) => a -> a -> v a

mk3 :: (Vector v a, Dim v ~ 3) => a -> a -> a -> v a

mk4 :: (Vector v a, Dim v ~ 4) => a -> a -> a -> a -> v a

mk5 :: (Vector v a, Dim v ~ 5) => a -> a -> a -> a -> a -> v a

mkN :: forall proxy v a. Vector v a => proxy (v a) -> Fn (Peano (Dim v)) a (v a)

N-ary constructor. Despite scary signature it's just N-ary function with additional type parameter which is used to fix type of vector being constructed. It could be used as:

v = mkN (Proxy :: Proxy (Int,Int,Int)) 1 2 3

or using TypeApplications syntax:

v = mkN (Proxy @ (Int,Int,Int)) 1 2 3

or if type of v is fixed elsewhere

v = mkN [v] 1 2 3

Continuation-based vectors

data ContVec n a

Vector represented as continuation. Alternative wording: it's Church encoded N-element vector.

empty :: ContVec 0 a

Create empty vector.

vector :: (Vector v a, Dim v ~ n) => ContVec n a -> v a

Convert continuation to the vector.

cvec :: (Vector v a, Dim v ~ n) => v a -> ContVec n a

Convert regular vector to continuation based one.

Functions

replicate :: Vector v a => a -> v a

Replicate value n times.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec2)
>>> replicate 1 :: Vec2 Int
fromList [1,1]

>>> replicate 2 :: (Double,Double,Double)
(2.0,2.0,2.0)

>>> import Data.Vector.Fixed.Boxed (Vec4)
>>> replicate "foo" :: Vec4 String
fromList ["foo","foo","foo","foo"]

replicateM :: (Vector v a, Applicative f) => f a -> f (v a)

Execute monadic action for every element of vector.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec2,Vec3)
>>> replicateM (Just 3) :: Maybe (Vec3 Int)
Just (fromList [3,3,3])
>>> replicateM (putStrLn "Hi!") :: IO (Vec2 ())
Hi!
Hi!
fromList [(),()]

generate :: Vector v a => (Int -> a) -> v a

Generate vector from function which maps element's index to its value.

Examples:

>>> import Data.Vector.Fixed.Unboxed (Vec4)
>>> generate (^2) :: Vec4 Int
fromList [0,1,4,9]

generateM :: (Applicative f, Vector v a) => (Int -> f a) -> f (v a)

Generate vector from monadic function which maps element's index to its value.

unfoldr :: Vector v a => (b -> (a, b)) -> b -> v a

Unfold vector.

basis :: (Vector v a, Num a) => Int -> v a

Unit vector along Nth axis. If index is larger than vector dimensions returns zero vector.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec3)
>>> basis 0 :: Vec3 Int
fromList [1,0,0]
>>> basis 1 :: Vec3 Int
fromList [0,1,0]
>>> basis 3 :: Vec3 Int
fromList [0,0,0]

Modifying vectors

Transformations

head :: (Vector v a, 1 <= Dim v) => v a -> a

First element of vector.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec3)
>>> let x = mk3 1 2 3 :: Vec3 Int
>>> head x
1

tail :: (Vector v a, Vector w a, Dim v ~ (Dim w + 1)) => v a -> w a

Tail of vector.

Examples:

>>> import Data.Complex
>>> tail (1,2,3) :: Complex Double
2.0 :+ 3.0

cons :: (Vector v a, Vector w a, Dim w ~ (Dim v + 1)) => a -> v a -> w a

Cons element to the vector

snoc :: (Vector v a, Vector w a, Dim w ~ (Dim v + 1)) => a -> v a -> w a

Append element to the vector

concat :: (Vector v a, Vector u a, Vector w a, (Dim v + Dim u) ~ Dim w, Peano (Dim v + Dim u) ~ Add (Peano (Dim v)) (Peano (Dim u))) => v a -> u a -> w a

reverse :: Vector v a => v a -> v a

Reverse order of elements in the vector

Indexing & lenses

(!) :: Vector v a => v a -> Int -> a

Retrieve vector's element at index. Generic implementation is O(n) but more efficient one is used when possible.

index :: (Vector v a, KnownNat k, (k + 1) <= Dim v) => v a -> proxy k -> a

Get element from vector at statically known index

set :: (Vector v a, KnownNat k, (k + 1) <= Dim v) => proxy k -> a -> v a -> v a

Set n'th element in the vector

element :: (Vector v a, Functor f) => Int -> (a -> f a) -> v a -> f (v a)

Twan van Laarhoven's lens for element of vector

elementTy :: (Vector v a, KnownNat k, (k + 1) <= Dim v, Functor f) => proxy k -> (a -> f a) -> v a -> f (v a)

Twan van Laarhoven's lens for element of vector with statically known index.

Comparison

eq :: (Vector v a, Eq a) => v a -> v a -> Bool

Test two vectors for equality.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec2)
>>> let v0 = basis 0 :: Vec2 Int
>>> let v1 = basis 1 :: Vec2 Int
>>> v0 `eq` v0
True
>>> v0 `eq` v1
False

ord :: (Vector v a, Ord a) => v a -> v a -> Ordering

Lexicographic ordering of two vectors.

Maps

map :: (Vector v a, Vector v b) => (a -> b) -> v a -> v b

Map over vector

mapM :: (Vector v a, Vector v b, Applicative f) => (a -> f b) -> v a -> f (v b)

Effectful map over vector.

mapM_ :: (Vector v a, Applicative f) => (a -> f b) -> v a -> f ()

Apply monadic action to each element of vector and ignore result.

imap :: (Vector v a, Vector v b) => (Int -> a -> b) -> v a -> v b

Apply function to every element of the vector and its index.

imapM :: (Vector v a, Vector v b, Applicative f) => (Int -> a -> f b) -> v a -> f (v b)

Apply monadic function to every element of the vector and its index.

imapM_ :: (Vector v a, Applicative f) => (Int -> a -> f b) -> v a -> f ()

Apply monadic function to every element of the vector and its index and discard result.

scanl :: (Vector v a, Vector w b, Dim w ~ (Dim v + 1)) => (b -> a -> b) -> b -> v a -> w b

Left scan over vector

scanl1 :: Vector v a => (a -> a -> a) -> v a -> v a

Left scan over vector

sequence :: (Vector v a, Vector v (f a), Applicative f) => v (f a) -> f (v a)

Evaluate every action in the vector from left to right.

sequence_ :: (Vector v (f a), Applicative f) => v (f a) -> f ()

Evaluate every action in the vector from left to right and ignore result

sequenceA :: (Vector v a, Vector v (f a), Applicative f) => v (f a) -> f (v a)

Analog of sequenceA from Traversable.

traverse :: (Vector v a, Vector v b, Applicative f) => (a -> f b) -> v a -> f (v b)

Analog of traverse from Traversable.

distribute :: (Vector v a, Vector v (f a), Functor f) => f (v a) -> v (f a)

collect :: (Vector v a, Vector v b, Vector v (f b), Functor f) => (a -> v b) -> f a -> v (f b)

Folding

foldl :: Vector v a => (b -> a -> b) -> b -> v a -> b

Left fold over vector

foldr :: Vector v a => (a -> b -> b) -> b -> v a -> b

Right fold over vector

foldl1 :: (Vector v a, 1 <= Dim v) => (a -> a -> a) -> v a -> a

Left fold over vector

fold :: (Vector v m, Monoid m) => v m -> m

Combine the elements of a structure using a monoid. Similar to fold

foldMap :: (Vector v a, Monoid m) => (a -> m) -> v a -> m

Map each element of the structure to a monoid, and combine the results. Similar to foldMap

ifoldl :: Vector v a => (b -> Int -> a -> b) -> b -> v a -> b

Left fold over vector. Function is applied to each element and its index.

ifoldr :: Vector v a => (Int -> a -> b -> b) -> b -> v a -> b

Right fold over vector

foldM :: (Vector v a, Monad m) => (b -> a -> m b) -> b -> v a -> m b

Monadic fold over vector.

ifoldM :: (Vector v a, Monad m) => (b -> Int -> a -> m b) -> b -> v a -> m b

Left monadic fold over vector. Function is applied to each element and its index.

Special folds

sum :: (Vector v a, Num a) => v a -> a

Sum all elements in the vector.

maximum :: (Vector v a, 1 <= Dim v, Ord a) => v a -> a

Maximal element of vector.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec3)
>>> let x = mk3 1 2 3 :: Vec3 Int
>>> maximum x
3

minimum :: (Vector v a, 1 <= Dim v, Ord a) => v a -> a

Minimal element of vector.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec3)
>>> let x = mk3 1 2 3 :: Vec3 Int
>>> minimum x
1

and :: Vector v Bool => v Bool -> Bool

Conjunction of all elements of a vector.

or :: Vector v Bool => v Bool -> Bool

Disjunction of all elements of a vector.

all :: Vector v a => (a -> Bool) -> v a -> Bool

Determines whether all elements of vector satisfy predicate.

any :: Vector v a => (a -> Bool) -> v a -> Bool

Determines whether any of element of vector satisfy predicate.

find :: Vector v a => (a -> Bool) -> v a -> Maybe a

The find function takes a predicate and a vector and returns the leftmost element of the vector matching the predicate, or Nothing if there is no such element.

Zips

zipWith :: (Vector v a, Vector v b, Vector v c) => (a -> b -> c) -> v a -> v b -> v c

Zip two vector together using function.

Examples:

>>> import Data.Vector.Fixed.Boxed (Vec3)
>>> let b0 = basis 0 :: Vec3 Int
>>> let b1 = basis 1 :: Vec3 Int
>>> let b2 = basis 2 :: Vec3 Int
>>> let vplus x y = zipWith (+) x y
>>> vplus b0 b1
fromList [1,1,0]
>>> vplus b0 b2
fromList [1,0,1]
>>> vplus b1 b2
fromList [0,1,1]

zipWith3 :: (Vector v a, Vector v b, Vector v c, Vector v d) => (a -> b -> c -> d) -> v a -> v b -> v c -> v d

Zip three vector together

zipWithM :: (Vector v a, Vector v b, Vector v c, Applicative f) => (a -> b -> f c) -> v a -> v b -> f (v c)

Zip two vector together using monadic function.

zipWithM_ :: (Vector v a, Vector v b, Applicative f) => (a -> b -> f c) -> v a -> v b -> f ()

Zip two vector elementwise using monadic function and discard result

izipWith :: (Vector v a, Vector v b, Vector v c) => (Int -> a -> b -> c) -> v a -> v b -> v c

Zip two vector together using function which takes element index as well.

izipWith3 :: (Vector v a, Vector v b, Vector v c, Vector v d) => (Int -> a -> b -> c -> d) -> v a -> v b -> v c -> v d

Zip three vector together

izipWithM :: (Vector v a, Vector v b, Vector v c, Applicative f) => (Int -> a -> b -> f c) -> v a -> v b -> f (v c)

Zip two vector together using monadic function which takes element index as well..

izipWithM_ :: (Vector v a, Vector v b, Vector v c, Applicative f, Vector v (f c)) => (Int -> a -> b -> f c) -> v a -> v b -> f ()

Zip two vector elementwise using monadic function and discard result

Storable methods

Default implementation of methods for Storable type class assumes that individual elements of vector are stored as N-element array.

defaultAlignemnt :: forall a v. Storable a => v a -> Int

Default implementation of alignment for Storable type class for fixed vectors.

defaultSizeOf :: forall a v. (Storable a, Vector v a) => v a -> Int

Default implementation of sizeOf for Storable type class for fixed vectors

defaultPeek :: (Storable a, Vector v a) => Ptr (v a) -> IO (v a)

Default implementation of peek for Storable type class for fixed vector

defaultPoke :: (Storable a, Vector v a) => Ptr (v a) -> v a -> IO ()

Default implementation of poke for Storable type class for fixed vector

NFData

defaultRnf :: (NFData a, Vector v a) => v a -> ()

Default implementation of rnf from NFData type class

Conversion

convert :: (Vector v a, Vector w a, Dim v ~ Dim w) => v a -> w a

Convert between different vector types

toList :: Vector v a => v a -> [a]

Convert vector to the list

fromList :: Vector v a => [a] -> v a

Create vector form list. Will throw error if list is shorter than resulting vector.

fromList' :: Vector v a => [a] -> v a

Create vector form list. Will throw error if list has different length from resulting vector.

fromListM :: Vector v a => [a] -> Maybe (v a)

Create vector form list. Will return Nothing if list has different length from resulting vector.

fromFoldable :: (Vector v a, Foldable f) => f a -> Maybe (v a)

Create vector from Foldable data type. Will return Nothing if data type different number of elements that resulting vector.

Data types

newtype VecList (n :: Nat) a

Type-based vector with statically known length parametrized by GHC's type naturals

Constructors

VecList (VecPeano (Peano n) a)

data VecPeano (n :: PeanoNum) a where

Standard GADT-based vector with statically known length parametrized by Peano numbers.

Constructors

Nil :: VecPeano 'Z a

Cons :: a -> VecPeano n a -> VecPeano ('S n) a

newtype Only a

Single-element tuple.

Constructors

Only a

data Empty a

Empty tuple.

Constructors

Empty

Tuple synonyms

type Tuple2 a = (a, a)

type Tuple3 a = (a, a, a)

type Tuple4 a = (a, a, a, a)

type Tuple5 a = (a, a, a, a, a)