Data.Vector.Fixed.Storable

Description

Storable-based unboxed vectors.

Immutable

data Vec (n :: Nat) a

Storable-based vector with fixed length

type Vec1 = Vec 1

type Vec2 = Vec 2

type Vec3 = Vec 3

type Vec4 = Vec 4

type Vec5 = Vec 5

Raw pointers

unsafeFromForeignPtr :: ForeignPtr a -> Vec n a

Construct vector from foreign pointer.

unsafeToForeignPtr :: Vec n a -> ForeignPtr a

Get underlying pointer. Data may not be modified through pointer.

unsafeWith :: (Ptr a -> IO b) -> Vec n a -> IO b

Mutable

newtype MVec (n :: Nat) s a

Storable-based mutable vector with fixed length

Constructors

MVec (ForeignPtr a)

Type classes

class Storable a

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)