Data.Vector.Fixed.Cont

Description

API for Church-encoded vectors. Implementation of function from Data.Vector.Fixed module uses these function internally in order to provide shortcut fusion.

Type-level numbers

data PeanoNum

Peano numbers. Since type level naturals don't support induction we have to convert type nats to Peano representation first and work with it,

Constructors

Z

S PeanoNum

type family Peano (n :: Nat) :: PeanoNum where ...

Convert type level natural to Peano representation

Equations

Peano 0 = Z

Peano n = S (Peano (n - 1))

type family Add (n :: PeanoNum) (m :: PeanoNum) :: PeanoNum where ...

Type family for sum of unary natural numbers.

Equations

Add Z n = n

Add (S n) k = S (Add n k)

N-ary functions

type family Fn (n :: PeanoNum) (a :: *) (b :: *) where ...

Type family for n-ary functions. n is number of parameters of type a and b is result type.

Equations

Fn Z a b = b

Fn (S n) a b = a -> Fn n a b

newtype Fun n a b

Newtype wrapper which is used to make Fn injective. It's also a reader monad.

Constructors

Fun

Fields

unFun :: Fn n a b

type Arity n = (ArityPeano (Peano n), KnownNat n, Peano (n + 1) ~ S (Peano n))

Type class for type level number for which we can defined operations over N-ary functions.

class ArityPeano n where

Type class for handling n-ary functions.

Minimal complete definition

accum, applyFun, applyFunM, reverseF, gunfoldF

Methods

accum

Arguments

:: (forall k. t (S k) -> a -> t k)
Fold function

-> (t Z -> b)
Extract result of fold

-> t n
Initial value

-> Fun n a b
Reduction function

Left fold over n elements exposed as n-ary function. These elements are supplied as arguments to the function.

applyFun

Arguments

:: (forall k. t (S k) -> (a, t k))
Get value to apply to function

-> t n
Initial value

-> (CVecPeano n a, t Z)

Apply all parameters to the function.

applyFunM

Arguments

:: Applicative f

=> (forall k. t (S k) -> (f a, t k))
Get value to apply to function

-> t n
Initial value

-> (f (CVecPeano n a), t Z)

Apply all parameters to the function using monadic actions. Note that for identity monad it's same as applyFun. Ignoring newtypes:

forall b. Fn n a b -> b  ~ ContVec n a

reverseF :: Fun n a b -> Fun n a b

Reverse order of parameters. It's implemented directly in type class since expressing it in terms of accum will require putting ArityPeano constraint on step funcion

gunfoldF :: Data a => (forall b x. Data b => c (b -> x) -> c x) -> T_gunfold c r a n -> c r

Worker function for gunfold

arity :: KnownNat n => proxy n -> Int

Arity of function.

apply

Arguments

:: Arity n

=> (forall k. t (S k) -> (a, t k))
Get value to apply to function

-> t (Peano n)
Initial value

-> ContVec n a
N-ary function

Apply all parameters to the function.

applyM

Arguments

:: (Applicative f, Arity n)

=> (forall k. t (S k) -> (f a, t k))
Get value to apply to function

-> t (Peano n)
Initial value

-> f (ContVec n a)

Apply all parameters to the function using applicative actions.

Combinators

constFun :: Fun n a b -> Fun (S n) a b

Prepend ignored parameter to function

curryFirst :: Fun (S n) a b -> a -> Fun n a b

Curry first parameter of n-ary function

uncurryFirst :: (a -> Fun n a b) -> Fun (S n) a b

Uncurry first parameter of n-ary function

curryLast :: ArityPeano n => Fun (S n) a b -> Fun n a (a -> b)

Curry last parameter of n-ary function

curryMany :: forall n k a b. ArityPeano n => Fun (Add n k) a b -> Fun n a (Fun k a b)

Curry n first parameters of n-ary function

apLast :: ArityPeano n => Fun (S n) a b -> a -> Fun n a b

Apply last parameter to function. Unlike apFun we need to traverse all parameters but last hence Arity constraint.

shuffleFun :: ArityPeano n => (b -> Fun n a r) -> Fun n a (b -> r)

Move function parameter to the result of N-ary function.

withFun :: (Fun n a b -> Fun n a b) -> Fun (S n) a b -> Fun (S n) a b

Recursive step for the function

Vector type class

type family Dim (v :: * -> *) :: Nat

Size of vector expressed as type-level natural.

class Arity (Dim v) => Vector v a where

Type class for vectors with fixed length. Instance should provide two functions: one to create vector and another for vector deconstruction. They must obey following law:

inspect v construct = v

For example instance for 2D vectors could be written as:

data V2 a = V2 a a

type instance V2 = 2
instance Vector V2 a where
  construct                = Fun V2
  inspect (V2 a b) (Fun f) = f a b

Minimal complete definition

construct, inspect

Methods

construct :: Fun (Peano (Dim v)) a (v a)

N-ary function for creation of vectors.

inspect :: v a -> Fun (Peano (Dim v)) a b -> b

Deconstruction of vector.

basicIndex :: v a -> Int -> a

Optional more efficient implementation of indexing. Shouldn't be used directly, use ! instead.

class (Vector (v n) a, Dim (v n) ~ n) => VectorN v n a

Vector parametrized by length. In ideal world it should be:

forall n. (Arity n, Vector (v n) a, Dim (v n) ~ n) => VectorN v a

Alas polymorphic constraints aren't allowed in haskell.

length :: forall v a. KnownNat (Dim v) => v a -> Int

Length of vector. Function doesn't evaluate its argument.

Vector as continuation

newtype ContVec n a

Vector represented as continuation. Alternative wording: it's Church encoded N-element vector.

Constructors

ContVec (forall r. Fun (Peano n) a r -> r)

newtype CVecPeano n a

Same as ContVec but its length is expressed as Peano number.

Constructors

CVecPeano (forall r. Fun n a r -> r)

consPeano :: a -> CVecPeano n a -> CVecPeano (S n) a

Cons values to the CVecPeano.

toContVec :: CVecPeano (Peano n) a -> ContVec n a

runContVec :: Fun (Peano n) a r -> ContVec n a -> r

Run continuation vector. It's same as inspect but with arguments flipped.

Construction of ContVec

cvec :: (Vector v a, Dim v ~ n) => v a -> ContVec n a

Convert regular vector to continuation based one.

fromList :: Arity n => [a] -> ContVec n a

Convert list to continuation-based vector. Will throw error if list is shorter than resulting vector.

fromList' :: forall n a. Arity n => [a] -> ContVec n a

Same as fromList bu throws error is list doesn't have same length as vector.

fromListM :: forall n a. Arity n => [a] -> Maybe (ContVec n a)

Convert list to continuation-based vector. Will fail with Nothing if list doesn't have right length.

toList :: Arity n => ContVec n a -> [a]

Convert vector to the list

replicate :: Arity n => a -> ContVec n a

Execute monadic action for every element of vector. Synonym for pure.

replicateM :: (Arity n, Applicative f) => f a -> f (ContVec n a)

Execute monadic action for every element of vector.

generate :: Arity n => (Int -> a) -> ContVec n a

Generate vector from function which maps element's index to its value.

generateM :: (Applicative f, Arity n) => (Int -> f a) -> f (ContVec n a)

Generate vector from monadic function which maps element's index to its value.

unfoldr :: Arity n => (b -> (a, b)) -> b -> ContVec n a

Unfold vector.

basis :: (Num a, Arity n) => Int -> ContVec n a

Unit vector along Nth axis.

Constructors

empty :: ContVec 0 a

Create empty vector.

cons :: Arity n => a -> ContVec n a -> ContVec (n + 1) a

O(1) Prepend element to vector

consV :: Arity n => ContVec 1 a -> ContVec n a -> ContVec (n + 1) a

Prepend single element vector to another vector.

snoc :: Arity n => a -> ContVec n a -> ContVec (n + 1) a

O(1) Append element to vector

concat :: (Arity n, Arity k, Arity (n + k), Peano (n + k) ~ Add (Peano n) (Peano k)) => ContVec n a -> ContVec k a -> ContVec (n + k) a

Concatenate vector

mk1 :: a -> ContVec 1 a

mk2 :: a -> a -> ContVec 2 a

mk3 :: a -> a -> a -> ContVec 3 a

mk4 :: a -> a -> a -> a -> ContVec 4 a

mk5 :: a -> a -> a -> a -> a -> ContVec 5 a

Transformations

map :: Arity n => (a -> b) -> ContVec n a -> ContVec n b

Map over vector. Synonym for fmap

imap :: Arity n => (Int -> a -> b) -> ContVec n a -> ContVec n b

Apply function to every element of the vector and its index.

mapM :: (Arity n, Applicative f) => (a -> f b) -> ContVec n a -> f (ContVec n b)

Effectful map over vector.

imapM :: (Arity n, Applicative f) => (Int -> a -> f b) -> ContVec n a -> f (ContVec n b)

Apply monadic function to every element of the vector and its index.

mapM_ :: (Arity n, Applicative f) => (a -> f b) -> ContVec n a -> f ()

Apply monadic action to each element of vector and ignore result.

imapM_ :: (Arity n, Applicative f) => (Int -> a -> f b) -> ContVec n a -> f ()

Apply monadic action to each element of vector and its index and ignore result.

scanl :: Arity n => (b -> a -> b) -> b -> ContVec n a -> ContVec (n + 1) b

Left scan over vector

scanl1 :: Arity n => (a -> a -> a) -> ContVec n a -> ContVec n a

Left scan over vector

sequence :: (Arity n, Applicative f) => ContVec n (f a) -> f (ContVec n a)

Evaluate every action in the vector from left to right.

sequence_ :: (Arity n, Applicative f) => ContVec n (f a) -> f ()

Evaluate every action in the vector from left to right and ignore result.

distribute :: (Functor f, Arity n) => f (ContVec n a) -> ContVec n (f a)

The dual of sequenceA

collect :: (Functor f, Arity n) => (a -> ContVec n b) -> f a -> ContVec n (f b)

tail :: Arity n => ContVec (n + 1) a -> ContVec n a

O(1) Tail of vector.

reverse :: Arity n => ContVec n a -> ContVec n a

Reverse order of elements in the vector

Zips

zipWith :: Arity n => (a -> b -> c) -> ContVec n a -> ContVec n b -> ContVec n c

Zip two vector together using function.

zipWith3 :: Arity n => (a -> b -> c -> d) -> ContVec n a -> ContVec n b -> ContVec n c -> ContVec n d

Zip three vectors together

izipWith :: Arity n => (Int -> a -> b -> c) -> ContVec n a -> ContVec n b -> ContVec n c

Zip two vector together using function which takes element index as well.

izipWith3 :: Arity n => (Int -> a -> b -> c -> d) -> ContVec n a -> ContVec n b -> ContVec n c -> ContVec n d

Zip three vectors together

zipWithM :: (Arity n, Applicative f) => (a -> b -> f c) -> ContVec n a -> ContVec n b -> f (ContVec n c)

Zip two vector together using monadic function.

zipWithM_ :: (Arity n, Applicative f) => (a -> b -> f c) -> ContVec n a -> ContVec n b -> f ()

izipWithM :: (Arity n, Applicative f) => (Int -> a -> b -> f c) -> ContVec n a -> ContVec n b -> f (ContVec n c)

Zip two vector together using monadic function which takes element index as well..

izipWithM_ :: (Arity n, Applicative f) => (Int -> a -> b -> f c) -> ContVec n a -> ContVec n b -> f ()

Getters

head :: (Arity n, 1 <= n) => ContVec n a -> a

Finalizer function for getting head of the vector.

index :: Arity n => Int -> ContVec n a -> a

O(n) Get value at specified index.

element :: (Arity n, Functor f) => Int -> (a -> f a) -> ContVec n a -> f (ContVec n a)

Twan van Laarhoven lens for continuation based vector

Vector construction

vector :: (Vector v a, Dim v ~ n) => ContVec n a -> v a

Convert continuation to the vector.

Folds

foldl :: Arity n => (b -> a -> b) -> b -> ContVec n a -> b

Left fold over continuation vector.

foldl1 :: (Arity n, 1 <= n) => (a -> a -> a) -> ContVec n a -> a

Left fold.

foldr :: Arity n => (a -> b -> b) -> b -> ContVec n a -> b

Right fold over continuation vector

ifoldl :: Arity n => (b -> Int -> a -> b) -> b -> ContVec n a -> b

Left fold over continuation vector.

ifoldr :: Arity n => (Int -> a -> b -> b) -> b -> ContVec n a -> b

Right fold over continuation vector

foldM :: (Arity n, Monad m) => (b -> a -> m b) -> b -> ContVec n a -> m b

Monadic left fold over continuation vector.

ifoldM :: (Arity n, Monad m) => (b -> Int -> a -> m b) -> b -> ContVec n a -> m b

Monadic left fold over continuation vector.

Special folds

sum :: (Num a, Arity n) => ContVec n a -> a

Sum all elements in the vector.

minimum :: (Ord a, Arity n, 1 <= n) => ContVec n a -> a

Minimal element of vector.

maximum :: (Ord a, Arity n, 1 <= n) => ContVec n a -> a

Maximal element of vector.

and :: Arity n => ContVec n Bool -> Bool

Conjunction of elements of a vector.

or :: Arity n => ContVec n Bool -> Bool

Disjunction of all elements of a vector.

all :: Arity n => (a -> Bool) -> ContVec n a -> Bool

Determines whether all elements of vector satisfy predicate.

any :: Arity n => (a -> Bool) -> ContVec n a -> Bool

Determines whether any of element of vector satisfy predicate.

find :: Arity n => (a -> Bool) -> ContVec n a -> Maybe a

The find function takes a predicate and a vector and returns the leftmost element of the vector matching the predicate, or Nothing if there is no such element.

Data.Data.Data

gfoldl :: forall c v a. (Vector v a, Data a) => (forall x y. Data x => c (x -> y) -> x -> c y) -> (forall x. x -> c x) -> v a -> c (v a)

Generic gfoldl which could work with any vector.

gunfold :: forall con c v a. (Vector v a, Data a) => (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> con -> c (v a)

Generic gunfoldl which could work with any vector. Since vector can only have one constructor argument for constructor is ignored.