Data.Vector.Fixed.Mutable

Description

Type classes for vectors which are implemented on top of the arrays and support in-place mutation. API is similar to one used in the vector package.

Mutable vectors

type Arity n = (ArityPeano (Peano n), KnownNat n, Peano (n + 1) ~ S (Peano n))

Type class for type level number for which we can defined operations over N-ary functions.

arity :: KnownNat n => proxy n -> Int

Arity of function.

type family Mutable (v :: * -> *) :: * -> * -> *

Mutable counterpart of fixed-length vector.

type family DimM (v :: * -> * -> *) :: Nat

Dimension for mutable vector.

class Arity (DimM v) => MVector v a where

Type class for mutable vectors.

Minimal complete definition

copy, move, new, unsafeRead, unsafeWrite

Methods

copy

Arguments

:: PrimMonad m

=> v (PrimState m) a
Target

-> v (PrimState m) a
Source

-> m ()

Copy vector. The two vectors may not overlap. Since vectors' length is encoded in the type there is no need in runtime checks.

move

Arguments

:: PrimMonad m

=> v (PrimState m) a
Target

-> v (PrimState m) a
Source

-> m ()

Copy vector. The two vectors may overlap. Since vectors' length is encoded in the type there is no need in runtime checks.

new :: PrimMonad m => m (v (PrimState m) a)

Allocate new vector

unsafeRead :: PrimMonad m => v (PrimState m) a -> Int -> m a

Read value at index without bound checks.

unsafeWrite :: PrimMonad m => v (PrimState m) a -> Int -> a -> m ()

Write value at index without bound checks.

lengthM :: forall v s a. Arity (DimM v) => v s a -> Int

Length of mutable vector. Function doesn't evaluate its argument.

read :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> m a

Read value at index with bound checks.

write :: (PrimMonad m, MVector v a) => v (PrimState m) a -> Int -> a -> m ()

Write value at index with bound checks.

clone :: (PrimMonad m, MVector v a) => v (PrimState m) a -> m (v (PrimState m) a)

Create copy of vector.

Immutable vectors

class (Dim v ~ DimM (Mutable v), MVector (Mutable v) a) => IVector v a where

Type class for immutable vectors

Minimal complete definition

unsafeFreeze, unsafeThaw, unsafeIndex

Methods

unsafeFreeze :: PrimMonad m => Mutable v (PrimState m) a -> m (v a)

Convert vector to immutable state. Mutable vector must not be modified afterwards.

unsafeThaw :: PrimMonad m => v a -> m (Mutable v (PrimState m) a)

Convert immutable vector to mutable. Immutable vector must not be used afterwards.

unsafeIndex :: v a -> Int -> a

Get element at specified index without bounds check.

index :: IVector v a => v a -> Int -> a

freeze :: (PrimMonad m, IVector v a) => Mutable v (PrimState m) a -> m (v a)

Safely convert mutable vector to immutable.

thaw :: (PrimMonad m, IVector v a) => v a -> m (Mutable v (PrimState m) a)

Safely convert immutable vector to mutable.

Vector API

constructVec :: forall v a. (Arity (Dim v), IVector v a) => Fun (Peano (Dim v)) a (v a)

Generic construct implementation for array-based vectors.

inspectVec :: forall v a b. (Arity (Dim v), IVector v a) => v a -> Fun (Peano (Dim v)) a b -> b

Generic inspect implementation for array-based vectors.